fix: Massive changes to the method of routing
Previously, the elements used the Link component's state attribute to
send the current cart to the next route and it worked, until it didn't.
The problem with this method is the state is saved in history, so when
going back to a previous route(in history timeline) the element used the
state it was saved in history instead of the state from the previous
element. In short, forwards was good and backwards was bad.

While trying to think of a working method, I thought about using state
in an element and then rendering a router as a child. While the method
would work, a new router would be rendered every state change, so it was
a no-go(thanks to nevz on TOP Discord for providing such insight).

The method I implemented(suggested by nevz) was to use an element as the
index page and put the state on there. Then have the other routes as
child routes to make use of useOutletState to pass it down to children(
basically lifting state up to accommodate every needed route). If the
page is reloaded, a new index element is loaded and new state is
created.